ch376-native: updated keyboard driver
1. fix handling of caps-lock
2. removed support for returning scancode/bitfields for hbios call

+++ Source/HBIOS/ch376-native/source-doc/keyboard/kyb_driver.h
@@ -7,6 +7,6 @@
 extern void     usb_kyb_init(const uint8_t dev_index) __sdcccall(1);
 extern uint8_t  usb_kyb_flush() __sdcccall(1);
 extern uint8_t  usb_kyb_status() __sdcccall(1);
-extern uint32_t usb_kyb_read();
+extern uint16_t usb_kyb_read();
 
 #endif
